refactored typestag
